Add MouseAnchor and SmartAnchor features
MouseAnchor causes the tooltip to anchor to and follow the mouse when on WorldFrame (so you can see it next to a mob, for example, rather than looking back to the bottom right corner of the screen where your eyes likely are not)
SmartAnchor causes this behavior to be relatively intelligent and not follow across unitframes, etc - instead choosing to anchor 10 pixels above the frame it's hovering over in a static position for readability. 

In the current implementation, SmartAnchor does nothing if enabled without MouseAnchor - perhaps it should just be one feature called Smart Anchor? I'll leave that up to you as you are the approving authority. Hope I did the menu stuff right, I think I did and it does work.